draft InputHelper
